separate used variables
