+++ composeApp/src/commonMain/kotlin/org/example/project/App.kt
@@ -14,7 +14,7 @@ import kmp_bottom_bar.composeapp.generated.resources.papers
 import org.example.core.bottombar.BottomBarIdentifier
 import org.example.tinyGlide.data.TinyGlideItem
 import org.jetbrains.compose.ui.tooling.preview.Preview
-import ui.TinyGlideBottomBar
+import org.example.tinyGlide.bottombar.TinyGlideBottomBar
 
 
 @Composable
