+++ tinyGlideBottomBar/src/commonMain/kotlin/org/example/tinyGlide/data/TinyGlideItem.kt
@@ -6,6 +6,7 @@ import androidx.compose.ui.unit.dp
 import org.example.core.bottombar.BottomBarIdentifier
 import org.example.core.bottombar.BottomBarItem
 import org.jetbrains.compose.resources.DrawableResource
+import kotlin.time.Duration
 
 data class TinyGlideItem(
     override val identifier: BottomBarIdentifier,
@@ -14,5 +15,9 @@ data class TinyGlideItem(
     override val size: Dp = 50.dp,
     override val title: String = "options",
     val radius: Dp = 10.dp,
-    val backgroundColor: Color = Color.Blue
+    val backgroundColor: Color = Color.Blue,
+    val onSelectItemSizeChangeFriction : Float = 1.2f,
+    val onSelectItemSizeChangeDurationMillis: Int = 300,
+    val itemSeparationSpace: Dp = 10.dp,
+
 ) : BottomBarItem()
